Fix right-frame google translation link.
The current code uses a series of calls to eventually open a new window with a header
Location change.  I believe that Google has blocked this for XSS attack reasons.
This change just makes the proper URL using plain javascript and then opens the window.
It's very clear what's happening with this new code, but I don't know enough project
history to say if this is an acceptable change or not.

Note that the 'translate sentence' in the reading pane pop-up is still broken,
because it uses a slightly different set of calls, and still relies on the Location
header change.